requirements.txt
aiohttp
beautifulsoup4>=4.12.0
google-cloud-pubsub>=2.20.0
google-cloud-firestore>=2.14.0
pytest
pytest-asyncio
requests>=2.31.0
